python/requirements_py3.txt
requests==2.14.2
tensorflow==1.7.0
numpy==1.14.2
tensorflow-serving-api-python3~=1.7.0
protobuf==3.5.2
